pothos-library: switched serialization to portable text archive
We have successfully tested the remote stuff mixing architectures arm + x64.
The serialization for object is now text archive because binary wasnt portable.
Also there was a fix for the remote proxy to store the lookup TID value as a fixed size integer,
otherwise the tid could get truncated if it was a size_t which varies across architecture.

+++ pothos-library/lib/Remote/RemoteProxy.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2013-2014 Josh Blum
+// Copyright (c) 2013-2015 Josh Blum
 // SPDX-License-Identifier: BSL-1.0
 
 #include "RemoteProxy.hpp"
@@ -9,6 +9,7 @@
 #include <iostream>
 #include <sstream>
 #include <thread>
+#include <cstdint>
 
 Pothos::ObjectKwargs RemoteProxyEnvironment::transact(const Pothos::ObjectKwargs &reqArgs_)
 {
@@ -18,7 +19,8 @@ Pothos::ObjectKwargs RemoteProxyEnvironment::transact(const Pothos::ObjectKwargs
     }
 
     //add the thread ID to the args
-    const auto tid = std::hash<std::thread::id>()(std::this_thread::get_id());
+    //tid must be a fixed size type so it doesn't get truncated through serialization
+    const auto tid = uint32_t(std::hash<std::thread::id>()(std::this_thread::get_id()));
     auto reqArgs = reqArgs_;
     reqArgs["tid"] = Pothos::Object(tid);
